Update shortcode-give.html.twig
Changed html template to match the ckeditor 5 button plugin html.
Button does not have a light gray or dark gray option so I set dark to be black and light to be gray as to emulate the same color options.
I also added logic for style to be set to "default" instead of "regular" as there was no differentiation in D7